+++ mocks/frontmatter-set-operation.go
@@ -9,7 +9,7 @@ import (
 )
 
 type FrontmatterSetOperation struct {
-	ExecuteStub        func(context.Context, string, string, string, string, string, string) error
+	ExecuteStub        func(context.Context, string, string, string, string, string, string, bool) error
 	executeMutex       sync.RWMutex
 	executeArgsForCall []struct {
 		arg1 context.Context
@@ -19,6 +19,7 @@ type FrontmatterSetOperation struct {
 		arg5 string
 		arg6 string
 		arg7 string
+		arg8 bool
 	}
 	executeReturns struct {
 		result1 error
@@ -30,7 +31,7 @@ type FrontmatterSetOperation struct {
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *FrontmatterSetOperation) Execute(arg1 context.Context, arg2 string, arg3 string, arg4 string, arg5 string, arg6 string, arg7 string) error {
+func (fake *FrontmatterSetOperation) Execute(arg1 context.Context, arg2 string, arg3 string, arg4 string, arg5 string, arg6 string, arg7 string, arg8 bool) error {
 	fake.executeMutex.Lock()
 	ret, specificReturn := fake.executeReturnsOnCall[len(fake.executeArgsForCall)]
 	fake.executeArgsForCall = append(fake.executeArgsForCall, struct {
@@ -41,13 +42,14 @@ func (fake *FrontmatterSetOperation) Execute(arg1 context.Context, arg2 string, 
 		arg5 string
 		arg6 string
 		arg7 string
-	}{arg1, arg2, arg3, arg4, arg5, arg6, arg7})
+		arg8 bool
+	}{arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8})
 	stub := fake.ExecuteStub
 	fakeReturns := fake.executeReturns
-	fake.recordInvocation("Execute", []interface{}{arg1, arg2, arg3, arg4, arg5, arg6, arg7})
+	fake.recordInvocation("Execute", []interface{}{arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8})
 	fake.executeMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3, arg4, arg5, arg6, arg7)
+		return stub(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8)
 	}
 	if specificReturn {
 		return ret.result1
@@ -61,17 +63,17 @@ func (fake *FrontmatterSetOperation) ExecuteCallCount() int {
 	return len(fake.executeArgsForCall)
 }
 
-func (fake *FrontmatterSetOperation) ExecuteCalls(stub func(context.Context, string, string, string, string, string, string) error) {
+func (fake *FrontmatterSetOperation) ExecuteCalls(stub func(context.Context, string, string, string, string, string, string, bool) error) {
 	fake.executeMutex.Lock()
 	defer fake.executeMutex.Unlock()
 	fake.ExecuteStub = stub
 }
 
-func (fake *FrontmatterSetOperation) ExecuteArgsForCall(i int) (context.Context, string, string, string, string, string, string) {
+func (fake *FrontmatterSetOperation) ExecuteArgsForCall(i int) (context.Context, string, string, string, string, string, string, bool) {
 	fake.executeMutex.RLock()
 	defer fake.executeMutex.RUnlock()
 	argsForCall := fake.executeArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5, argsForCall.arg6, argsForCall.arg7
+	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5, argsForCall.arg6, argsForCall.arg7, argsForCall.arg8
 }
 
 func (fake *FrontmatterSetOperation) ExecuteReturns(result1 error) {